change(api): use new consolidated routes for api spec and docs
To have a consistent and more intuitive way of accessing the OpenAPI
specification and the Swagger UI as a documentation of this API, accross
all base projects we will adopt these new routes on all projects.

To be still compatible with a frontend that is already consuming the old
routes, this change also adds redirects of the previous urls to the new
ones. This redirects will be removed in the future.